groblad_report --svalan: bug fix
Compensate for what is probably a change in Artportalen
behavior in early June: that the lack of trailing columns
does not count as empty trailing columns, but as some kind
of parse error.

Simply don't include the many fields (in the heading line)
which are never filled in, anyway.

In the future, we could also drop some always-empty fields
to the left.